converting indents to tabs, adding final comments
Added an alert to main.lua to detect the API key not being there and
direct people to the website to get their own.
Fixed bug in currentConditions where if the weather failed to load it
would generate an error.
Fixed a bug when the forecast data has a long text for the current
conditions description.